feat: Implement certificate generation and verification endpoints
- Added generate.controller.js for generating certificates with validation, logging, and error handling.
- Added verify.controller.js for verifying issued certificates with signature validation and logging.
- Enhanced contact.controller.js with Sentry logging for contact form submissions and email sending.
- Improved event.controller.js with detailed logging and error handling for event creation, editing, deletion, and fetching.
